ondevice-ai/requirements.txt
grpcio
grpcio-tools
flask
httpx
msgpack
numpy
PyYAML
pytest
